feat: port FFT-based BPM/signature/offset analysis to C++; fractional BPM
Replace the RMS-envelope autocorrelation tempo estimator in BpmAnalyzer with the
validated DSP pipeline (new bpm_dsp/fft sources, compiled via a CONFIGURE_DEPENDS
glob). analyze() now resamples decoded mono to 44.1 kHz, runs analyzeBpmDsp(), and
returns fractional BPM, first-downbeat offset, and detected beats-per-bar; BpmResult
gains a beatsPerBar field and nativeAnalyzeBpm returns a 3-element double array.

Propagate fractional BPM end to end: AudioEngine, ClickGenerator, and the JNI bridge
use double, and ClickGenerator accumulates a fractional beat interval so non-integer
tempos do not drift. TrackEntity.bpm/detectedBpm become Double (DB version 3 -> 4),
PlaybackViewModel applies the detected signature, and the UI rounds BPM for display.

+++ app/src/main/cpp/AudioEngine.h
@@ -71,7 +71,7 @@ public:
      */
     void seekTo(int64_t positionMs);
 
-    void setBpm(int bpm);
+    void setBpm(double bpm);
     void setBeatsPerBar(int beatsPerBar);
     void setTrackVolume(float volume);
     void setClickVolume(float volume);
@@ -94,7 +94,7 @@ private:
 
     std::shared_ptr<oboe::AudioStream> mStream;
 
-    std::atomic<int>   mBpm{120};
+    std::atomic<double> mBpm{120.0};
     std::atomic<int>   mBeatsPerBar{4};
     std::atomic<float> mTrackVolume{1.0f};
     std::atomic<float> mClickVolume{1.0f};
@@ -111,7 +111,7 @@ private:
 
     // Last values forwarded to mClickGenerator; 0 forces configure() on the
     // first onAudioReady call.
-    int mLastBpm{0};
+    double mLastBpm{0.0};
     int mLastBeatsPerBar{0};
 
     /**
