+++ app/src/main/java/com/example/tempoz/data/TrackEntity.kt
@@ -8,9 +8,9 @@ import androidx.room.PrimaryKey
 data class TrackEntity(
     @PrimaryKey val uri: String,
     @ColumnInfo(name = "display_name") val displayName: String,
-    @ColumnInfo(name = "bpm") val bpm: Int,
+    @ColumnInfo(name = "bpm") val bpm: Double,
     @ColumnInfo(name = "beats_per_bar") val beatsPerBar: Int,
     @ColumnInfo(name = "last_used_ms") val lastUsedMs: Long,
-    @ColumnInfo(name = "detected_bpm") val detectedBpm: Int? = null,
+    @ColumnInfo(name = "detected_bpm") val detectedBpm: Double? = null,
     @ColumnInfo(name = "beat_offset_frames") val beatOffsetFrames: Long? = null
 )
